Color Fade
This is working color fade. There is still some testing that needs to be done and removal of print statements. See TODO and remove in future. Also would be a good idea to add member variable to allow the user to select startup mode (Fade or not) at construction of neoPixel object.